implementing authorization

+++ server/routes/posts.js
@@ -75,9 +75,10 @@ router.post(
       const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId
     });
-    console.log('im here1');
+
       post.save().then(createdPost => {
         res.status(201).json({
           message: "Post added successfully",
@@ -102,18 +103,25 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
         title: req.body.title,
         content: req.body.content,
         imagePath: imagePath,
+        creator: req.userData.userId
     });
 
-    Post.updateOne({_id: req.params.id}, post).then(updatedPost => {
-        res.status(200).json({
-            message: 'Post updated sucessfully'
-        })
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if(result.nModified > 0){
+            res.status(200).json({message: 'Post updated sucessfully'});
+        } else {
+            res.status(401).json({message: 'Not authorized'});
+        }
     });
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        res.status(200).json({message: 'post deleted!'});
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId }).then(result => {
+        if(result.n > 0){
+            res.status(200).json({message: 'post deleted!'});
+        } else {
+            res.status(401).json({message: 'Not authorized'});
+        }
     });
 });
